# api/openapi.yaml
openapi: 3.1.0
info:
  title: CarMarket API
  version: 0.1.0
  description: India used car marketplace REST API
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /reviews:
    get:
      summary: List reviews for user or dealer store
      tags: [reviews]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: dealer_store_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reviews
    post:
      summary: Create review (requires accepted inquiry)
      tags: [reviews]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /reviews/{review_id}/reply:
    post:
      summary: Seller reply to review
      tags: [reviews]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Replied
  /reviews/{review_id}/report:
    post:
      summary: Report a review
      tags: [reviews]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reported
...
  /saved-searches:
    get:
      summary: List saved searches
      tags: [saved-searches]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved searches
    post:
      summary: Create saved search
      tags: [saved-searches]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /saved-searches/{search_id}:
    get:
      summary: Get saved search
      tags: [saved-searches]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Detail
    patch:
      summary: Update saved search
      tags: [saved-searches]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete saved search
      tags: [saved-searches]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT